Define continents and straighten out initialization
Previously, the concept of "ECMAScript implementation" initialization was highly implicit, and it was not clear that e.g. a browser would contain multiple "ECMAScript implementations", each with their own set of job queues. We make that explicit here by defining a "continent", per previous discussions in e.g. https://esdiscuss.org/topic/event-loops-in-navigated-away-from-windows and https://esdiscuss.org/topic/holy-mixed-metaphors-batman-was-event-loops-in-navigated-away-from-windows.

The previous "ECMAScript Initialization()" abstract operation was intended for use in initializing what we now call continents. However, it was ill-suited for its purpose, as each continent does not need a realm implicitly created, and in fact most continents will have more than one realm. As such, we leave continent creation up to the host environment, and merge "ECMAScript Initialization()" with InitializeHostDefinedRealm(). Host environments are now able to call InitializeHostDefinedRealm(global, thisValue) to create new realms within their own continents.

Since a continent now contains a set of job queues, it is now clearer (with the help of an additional non-normative note) that when doing EnqueueJob, a host should specify which continent they are working in.